Change: Move OSP connection funcs to manage_osp
The functions osp_connect_with_data and osp_scanner_connect are also
moved to manage_osp.
